Add files via upload
A simple autoclicker script that allows for easy changes of assigned keys and on/off toggles

Taken from a tutorial by Nitratine at https://nitratine.net/blog/post/python-auto-clicker/ with my own changes to the code. 
Added a simple Tkinter GUI that provides a brief explanation of the program and the source code link. 

- Developmental, my first program (ish since i did copy half of it ) on github so if anything is wrong let me know!